Skip Soft delete using move API (#1322)
* Adding new validator for move operation

* Updating some documentations

* Add config flag and use move opertation for object rename

* Adding unit test for configuration option

* Use move for folder renames

* Add Skeleton implementation of move operation for Vineer Client

* update apiary client version

* Adding Skelton implementation of Move operation in APIary client

* Implementing moveInternal method

* Adding unit tests

* Adding new validator for move operation

* Updating some documentations

* Add config flag and use move opertation for object rename

* Adding unit test for configuration option

* Use move for folder renames

* Adding integration test for rename operation with move

* Add implementation of moveInternal for gRPC

* Add unit test for Vineer(gRPC) client implementation

* Minor comment cleanups

* Minor comment cleanups

* Fixed tests to be conistent with current set up and made move_succeedsMultipleObjects more deterministic by sorting the response array. Enable integration test for gRPC.

* Updated comments

* Fixed maven spotless issues

* Fix(test): Create missing unique bucket for testRenameWithMoveOperation.

* Revert the integration test changes.The project is not whitelisted for the move API, so the integration test changes will fail.Will add back once the project is whitelisted.

* Updated move implementation to include a generation check for the source. Additionally, the implementation now handles the 'file not found' scenario without throwing an exception. This modification supports a retry mechanism, consistent with the approach used for deleteObjects.

* Added unit test for file not found and source generation check. Added basic implementation of move in InMemoryGoogleCloudStorage. Some additional maven spotless chnages

* Adding back exception on File Not found. We will implement single source,single dst api for gcsio layer.

* Addressed the comments in the PR

* Resolved nit comments

* Add back the integration test

* Ignore the integration tests as move not enabled for the project. Tested locally with  project.

---------

+++ gcsio/src/main/java/com/google/cloud/hadoop/gcsio/GoogleCloudStorageFileSystemImpl.java
@@ -55,6 +55,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.Comparator;
+import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
@@ -614,14 +615,25 @@ public class GoogleCloudStorageFileSystemImpl implements GoogleCloudStorageFileS
           StorageResourceId.fromUriPath(
               dst, /* allowEmptyObjectName= */ true, /* generationId= */ 0L);
 
-      gcs.copy(ImmutableMap.of(srcResourceId, dstResourceId));
+      if (this.options.getCloudStorageOptions().isMoveOperationEnabled()
+          && srcResourceId.getBucketName().equals(dstResourceId.getBucketName())) {
+        gcs.move(
+            ImmutableMap.of(
+                new StorageResourceId(
+                    srcInfo.getItemInfo().getBucketName(),
+                    srcInfo.getItemInfo().getObjectName(),
+                    srcInfo.getItemInfo().getContentGeneration()),
+                dstResourceId));
+      } else {
+        gcs.copy(ImmutableMap.of(srcResourceId, dstResourceId));
 
-      gcs.deleteObjects(
-          ImmutableList.of(
-              new StorageResourceId(
-                  srcInfo.getItemInfo().getBucketName(),
-                  srcInfo.getItemInfo().getObjectName(),
-                  srcInfo.getItemInfo().getContentGeneration())));
+        gcs.deleteObjects(
+            ImmutableList.of(
+                new StorageResourceId(
+                    srcInfo.getItemInfo().getBucketName(),
+                    srcInfo.getItemInfo().getObjectName(),
+                    srcInfo.getItemInfo().getContentGeneration())));
+      }
     }
 
     repairImplicitDirectory(srcParentInfoFuture);
@@ -764,6 +776,29 @@ public class GoogleCloudStorageFileSystemImpl implements GoogleCloudStorageFileS
       }
     }
 
+    StorageResourceId srcResourceId =
+        StorageResourceId.fromUriPath(src, /* allowEmptyObjectName= */ true);
+    StorageResourceId dstResourceId =
+        StorageResourceId.fromUriPath(
+            dst, /* allowEmptyObjectName= */ true, /* generationId= */ 0L);
+    if (this.options.getCloudStorageOptions().isMoveOperationEnabled()
+        && srcResourceId.getBucketName().equals(dstResourceId.getBucketName())) {
+
+      // First, move all items except marker items
+      moveInternal(srcToDstItemNames);
+      // Finally, move marker items (if any) to mark rename operation success
+      moveInternal(srcToDstMarkerItemNames);
+
+      if (srcInfo.getItemInfo().isBucket()) {
+        deleteBucket(Collections.singletonList(srcInfo));
+      } else {
+        // If src is a directory then srcItemInfos does not contain its own name,
+        // we delete item separately in the list.
+        deleteObjects(Collections.singletonList(srcInfo));
+      }
+      return;
+    }
+
     // First, copy all items except marker items
     copyInternal(srcToDstItemNames);
     // Finally, copy marker items (if any) to mark rename operation success
@@ -813,6 +848,27 @@ public class GoogleCloudStorageFileSystemImpl implements GoogleCloudStorageFileS
 
     // Perform copy.
     gcs.copy(srcBucketName, srcObjectNames, dstBucketName, dstObjectNames);
+  }
+
+  /** Moves items in given map that maps source items to destination items. */
+  private void moveInternal(Map<FileInfo, URI> srcToDstItemNames) throws IOException {
+    if (srcToDstItemNames.isEmpty()) {
+      return;
+    }
+
+    Map<StorageResourceId, StorageResourceId> sourceToDestinationObjectsMap = new HashMap<>();
+
+    // Prepare list of items to move.
+    for (Map.Entry<FileInfo, URI> srcToDstItemName : srcToDstItemNames.entrySet()) {
+      StorageResourceId srcResourceId = srcToDstItemName.getKey().getItemInfo().getResourceId();
+
+      StorageResourceId dstResourceId =
+          StorageResourceId.fromUriPath(srcToDstItemName.getValue(), true);
+      sourceToDestinationObjectsMap.put(srcResourceId, dstResourceId);
+    }
+
+    // Perform move.
+    gcs.move(sourceToDestinationObjectsMap);
   }
 
   /**
